+++ CHANGELOG.md
@@ -23,6 +23,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 * Domain Reload Roslyn analyzer
 * New Area category for IterationTime issues
 * Code/Domain Reload view, showing any issues raised by the Domain Reload Roslyn analyzer
+* Diagnostics for unsupported APIs on WebGL target
 
 ### Changed
 * Improved issue description and suggestion strings, to make them clearer and more consistent